requirements.txt
openai
gradio
requests
python-dotenv
bs4
pypdf2
numpy
scikit-learn
faiss-cpu


mcp
langchain
langchain-mcp-adapters
langchain-google-genai
langchain-ollama
langchain-core
google-adk

